+++ pycontest.egg-info/PKG-INFO
@@ -1,0 +1,31 @@
+Metadata-Version: 1.2
+Name: pycontest
+Version: 0.1.dev0
+Summary: pycontest: project used during PyCon tutorial
+Home-page: UNKNOWN
+Author: Dorota Jarecka & Anna Jaruga
+Author-email: djarecka@gmail.com
+Maintainer: Dorota Jatecka & Anna Jaruga
+Maintainer-email: djarecka@gmail.com
+License: MIT
+Description: 
+        pycontest
+        ========
+        
+        License
+        =======
+        ``pycontest`` is licensed under the terms of the MIT license. See the file
+        "LICENSE" for information on the history of this software, terms & conditions
+        for usage, and a DISCLAIMER OF ALL WARRANTIES.
+        
+Platform: OS Independent
+Classifier: Development Status :: 3 - Alpha
+Classifier: Environment :: Console
+Classifier: Intended Audience :: Science/Research
+Classifier: License :: OSI Approved :: MIT License
+Classifier: Operating System :: OS Independent
+Classifier: Programming Language :: Python
+Classifier: Topic :: Scientific/Engineering
+Requires: numpy
+Requires: scipy
+Requires: matplotlib
