+++ pycontest.egg-info/SOURCES.txt
@@ -1,0 +1,22 @@
+README.md
+setup.py
+pycontest/__init__.py
+pycontest/elastic_collisions.py
+pycontest/movies.py
+pycontest/simulation.py
+pycontest/transport.py
+pycontest/utils.py
+pycontest/version.py
+pycontest.egg-info/PKG-INFO
+pycontest.egg-info/SOURCES.txt
+pycontest.egg-info/dependency_links.txt
+pycontest.egg-info/requires.txt
+pycontest.egg-info/top_level.txt
+tests/__init__.py
+tests/test_elastic_collisions_1d.py
+tests/test_elastic_collisions_2d.py
+tests/test_hypothesis.py
+tests/test_regression.py
+tests/test_simulation.py
+tests/test_simulation_fixture.py
+tests/test_transport.py
