+++ tests/test_elastic_collisions_2d.py
@@ -15,9 +15,25 @@ def test_collision_2d_1():
 
 # you cannow do the same for y components
 def test_collision_2d_2():
-    pass
+    v1_f, v2_f = ec.collision_2d(v1=[0, 1], v2=[0, -2], r12=[0, 1], m1=1, m2=1)
+    assert (v1_f == [0, -2]).all()
+    assert (v2_f == [0, 1]).all()
+
 
 # you can also check the energy and momentum
+
+def test_collision_2d_en():
+    v1_i, v2_i = [3, 1], [1, -2]
+    m1, m2 = 1, 2
+    v1_f, v2_f = ec.collision_2d(v1=v1_i, v2=v2_i, r12=[0, 1], m1=m1, m2=m2)
+    assert E_kin([v1_i,v2_i], [m1, m2]) == E_kin([v1_f,v2_f], [m1, m2])
+
+
+def test_collision_2d_mom():
+    v1_i, v2_i = [3, 1], [1, -2]
+    m1, m2 = 1, 2
+    v1_f, v2_f = ec.collision_2d(v1=v1_i, v2=v2_i, r12=[0, 1], m1=m1, m2=m2)
+    assert (momentum([v1_i,v2_i], [m1, m2]) == momentum([v1_f, v2_f], [m1, m2])).all()
