+++ tests/test_simulation.py
@@ -5,12 +5,11 @@ import numpy as np
 import pytest
 
 
-
-def test_simulation_1():
+@pytest.mark.parametrize("dt", [0.5, 0.1, 0.01])
+def test_simulation_1(dt):
     # initial condition and simulation parameters
     domain = ([-2, 12], [0, 3])
     t_max = 6
-    dt = 0.5
     loc_0 = np.array([[0, 1.5],[10, 1.5]])
     vel_0 = np.array([[1, 0], [-1, 0]])
     radius = 1
@@ -34,3 +33,18 @@ def test_simulation_1():
     assert vel[1][0] == 1
     # Y component should stay the same
     assert (vel[0][1], vel[1][1]) == (vel_0[0][1], vel_0[1][1])
+
+
+def test_simulation_1_fail():
+    # initial condition and simulation parameters
+    domain = ([-2, 12], [0, 3])
+    dt = 1.
+    t_max = 12
+    loc_0 = np.array([[0, 1.5],[10, 1.5]])
+    vel_0 = np.array([[1, 0], [-1, 0]])
+    radius = 1
+    mass = [1, 1]
+    # checking if exception is raised properly
+    with pytest.raises(Exception) as excinfo:
+        loc, vel = sim.simulation(t_max, dt, mass, radius, loc_0, vel_0, domain)
+    assert "two particles are exactly in the same place" in str(excinfo.value)
